Finalize stateful `Heap` example
I've decided not to add explicit syntax support for stateful
testing on release. This should be done at a later date, so
I've added a very manual test of much the same thing instead.
This should be enough to illustrate the point that Supposition.jl
is well suited to generating arbitrary modifications of some
user-defined struct. Syntax sugar can come later.