Generate at least 5 elements in stateful docs
For educational purposes, it's better to show at least some elements
being generated, to showcase that what we're doing actually works.
However, it's not that great to overload a reader with the optional
keyword arguments while introducing an entirely different concept,
so let's hide the real invocation and retcon it later on to not mess
up the property test.