Finalize docalignment section
Expanding on the `sincosd` example, we can introduce how to handle
complicated consequences of computer numerics and how to test
these more general properties in a way that checks some guarantees
we'd expect to hold.

By using `patch` to show the change in docstring, we can better
showcase what has been added as a result of which discovery about
the function we're testing.