Implement basic history caching
The most basic history checking keeps track of one most recent
counterexample from a `@check` per

 * unique function name of the given property
 * unique argument types to the property
 * unique custom description passed to `@check`

Notably, this does NOT track the exact function or source location
of the invocation of `@check`. The reasoning for this is that
there may be a bug in the property or the property may move around,
in which case it would be bad if the cache got invalidated.

There are two types for managing these counterexamples in this commit,
`DirectoryDB` and `NoRecordDB`, both subtyping `ExampleDB`.
`ExampleDB` is the overarching interface abstract type. Types
implementing this interface must implement three  functions:

 * `records(::ExampleDB)`: Return a collection with all the keys
    managed by the DB
 * `record!(::ExampleDB, key, data)`: Stores all information required
    to replay a counterexample.
 * `retrieve(::ExampleDB, key)`: Retrieve a previously stored
    counterexample as a `Some`. Returns `nothing` if no data has been
    stored for this key.

`DirectoryDB` stores its data as a files on disk, one file per counter-
example. By default, the directory where counterexamples are stored
is in `SuppositionDB` under the `test` directory. This is customizable
on a per-@check basis. If requested, this can later be expanded to a
customizable global default, so that it doesn't have to be changed
at every invocation of `@check`.

`NoRecordDB` is special, in that it neither stores nor ever retrieves
counterexamples - hence the name, no record will be made. This is
intended as a "don't store counterexamples" variant, in case this
is desired.