Add test that error shrinking works
This used to work, right up until failure recording to a database
was added. By mistake, it was forgotten to actually record the failure,
instead trying to record the testset to itself. In addition, there was
a bug in the way targeting and error shrinking interacted with each
other. If we have an error, trying to hillclimb the error doesn't
make sense because we're primarily interested in finding the smallest
example that reproduces the error. Whether the example would
otherwise be a good counterexample (by growing the choices)
doesn't really matter anymore, since we can't complete the test anyway.

As a sideeffect, this makes error shrinking much faster!