Allow specifying the number of `tries` `example` should use
Previously, `example` just tried to `produce` an element from
the given `Possibility` 100_000 times before giving up.
With this patch, it's possible to have the function try a variable
number of times. Most people won't have the need to change this though,
as not being able to produce an element after 100_000 tries really
is indicative of a problem with the given `Possibility`.