Ensure default RNG is also reproducible
By seeding the default RNG from the (by default, hardware seeded)
given RNG, we can ensure that even calls to `rand` that aren't aware
of Supposition.jl are reproducible when replaying previous failures.

This is crucial, because it means that the explicitly given RNG is not
required to be passed around, so users don't have to change any
interfaces in order for Supposition.jl to fuzz their use of randomness
if they're not already explicitly handling RNG state.

There are some subtleties involved though - if a to-be-fuzzed code
uses bare `Xoshiro()`, this won't work anymore because that
constructor by default queries the `RandomDevice` (i.e. the hardware
RNG) for seeding. This is completely seperate from both the RNG
given to `@check`, as well as the default RNG, and so lies outside
of the control of Supposition.jl.