Add basic support for recursive generation
This adds the `Recursive` type, which is used for recursive
generation of objects by means of a base case `Possibility` and a
function taking that `Possibility`, returning a new `Possibility`
which produces wrapped in some other object.

The recursion factor is limited through the keyword argument
`max_layers`, indicating the maximum number of wrappings allowed
to be undertaken through recursion. This needs to be at least `1`,
so that at least the base case can always be produced.