Add Inf/NaN disabling to Floats
This adds two boolean keywords to `Floats`, `nans` and `infs`. The former
controls whether `NaN`s (signaling & quiet) can be generated, the latter
whether `Inf`s (positive and negative) can be generated.

I'm not too happy with how this shrinks, the construction I've implemented
here is just what I've come up with as a quick & dirty hack to make
it work at all. This definitely needs to be revisited in the future,
perhaps with custom/better shrinking passes that specifically target
floating point.