Ensure Data.Vectors can generate vectors close to the maximum size
Previously, we passed around the raw `Vector{UInt64}` of choices
and serialized that to disk. Now, we pass around an explicit
`Attempt` struct and serialize that instead, which holds all information
needed to reliably replay one `TestCase`.

Fixes #4

The way this used to work (hardcoding a probability) led to pretty much
all generated vectors having length <= 100. This commit smartly selects
a probability to achieve a desired average length, which is biased
to shorter lengths based on the average of a beta distribution,
steered through the "generation" of a testcase. The longer an invocation
of `@check` runs, the higher the generation. Earlier generations
produce shorter lengths, higher generations produce higher lengths.