Improve performance of single-arg `example`
This was previously refactored when `Vectors` was made smarter about
its distribution, but that came with a drop in performance.
This changes `example` again, so that instead of first generating
all possible results and then picking one, draw the generation
of the target _first_, and then directly use that as the result.