Add tests for error caching
This was previously untested; adding error caching is a good
opportunity to increase coverage too.

When shrinking an error, if another error is encountered, we need
to remember its location so that the user is only notified of each
distinct error once. "Distinct" here means either

 1) A different error type - if the types are different, the path to
    `throw`ing that error must have been different. This used to look
    at `==` of the thrown objects themselves, not their types,
    but that didn't play nice with custom errors with rich metadata.
 2) A different source location. If the error originates from a different
    location, the paths to getting that error must have been different
    too. This is likely a bit more conservative than necessary (it can
    happen that two errors from different locations "mean" the same thing
    to the program), but since we care about inputs exploring error
    paths, they should be treated as distinct errors too.

There is also a hook for figuring out which stackframe to report for
caching when handling an error, `find_user_error_frame`. This currently
only specializes for `ErrorException`, since that is commonly thrown
from the `error` function. Reporting the location of that function
instead of where it's called from as the "source" of the error
is more likely to be helpful in distinguishing different classes of
errors.