Add event tracking
This adds user-configurable event tracking through `event!`.

The advantage compared to logging through e.g. `@info` is that
the event tracking is silent, and only the events that occured in the final,
most minimal example are shown to the user. This can make it much easier
to debug certain behaviors, where both the minimal input as well
as some metadata encountered inside of the function is useful
for diagnosing an issue.