Ensure `@check` on an inline-defined function also respects RNG
Previously, this unintentionally didn't seed the default RNG from the
stream controlled by Supposition.jl, leading to the odd case that
Supposition.jl was able to find a counterexample, but then didn't
report that RNG correctly back when replaying the failure. This
fixes that, by simply seeding the default RNG in this form too.